style(JSON-RPC,starknet_client,storage): move tests under the module they are testing
