+++ crates/papyrus_execution/src/state_reader.rs
@@ -1,3 +1,7 @@
+#[cfg(test)]
+#[path = "state_reader_test.rs"]
+mod state_reader_test;
+
 use blockifier::execution::contract_class::ContractClass as BlockifierContractClass;
 use blockifier::state::errors::StateError;
 use blockifier::state::state_api::{StateReader as BlockifierStateReader, StateResult};
