+++ crates/papyrus_gateway/src/lib.rs
@@ -9,8 +9,6 @@ mod test_utils;
 mod v0_3_0;
 mod v0_4_0;
 mod version_config;
-#[cfg(test)]
-mod version_config_test;
 
 use std::collections::BTreeMap;
 use std::fmt::Display;
