+++ crates/papyrus_gateway/src/v0_3_0/broadcasted_transaction_test.rs
@@ -23,7 +23,8 @@ use starknet_api::transaction::{
 use starknet_client::writer::objects::transaction::DeprecatedContractClass;
 use test_utils::{auto_impl_get_test_instance, get_rng, GetTestInstance};
 
-use super::broadcasted_transaction::{
+use super::super::state::ContractClass;
+use super::{
     BroadcastedDeclareTransaction,
     BroadcastedDeclareV1Transaction,
     BroadcastedDeclareV2Transaction,
@@ -31,7 +32,6 @@ use super::broadcasted_transaction::{
     BroadcastedInvokeTransaction,
     BroadcastedTransaction,
 };
-use super::state::ContractClass;
 use crate::test_utils::{get_starknet_spec_api_schema_for_components, SpecFile};
 use crate::version_config::VERSION_0_3;
