+++ crates/papyrus_gateway/src/v0_3_0/mod.rs
@@ -1,11 +1,7 @@
 pub mod api;
 pub mod block;
 pub mod broadcasted_transaction;
-#[cfg(test)]
-mod broadcasted_transaction_test;
 pub mod deprecated_contract_class;
 pub mod error;
 pub mod state;
 pub mod transaction;
-#[cfg(test)]
-mod transaction_test;
