+++ crates/papyrus_gateway/src/v0_4_0/broadcasted_transaction_test.rs
@@ -17,13 +17,13 @@ use starknet_api::transaction::{Fee, TransactionSignature};
 use starknet_client::writer::objects::transaction::DeprecatedContractClass;
 use test_utils::{auto_impl_get_test_instance, get_number_of_variants, get_rng, GetTestInstance};
 
-use super::broadcasted_transaction::{
+use super::super::state::{ContractClass, EntryPointByType};
+use super::{
     BroadcastedDeclareTransaction,
     BroadcastedDeclareV1Transaction,
     BroadcastedDeclareV2Transaction,
     DeclareType,
 };
-use super::state::{ContractClass, EntryPointByType};
 use crate::test_utils::{get_starknet_spec_api_schema_for_components, SpecFile};
 use crate::version_config::VERSION_0_4;
