+++ crates/papyrus_gateway/src/v0_4_0/mod.rs
@@ -1,13 +1,9 @@
 pub mod api;
 pub mod block;
 pub mod broadcasted_transaction;
-#[cfg(test)]
-mod broadcasted_transaction_test;
 pub mod deprecated_contract_class;
 pub mod error;
 pub mod state;
 pub mod transaction;
-#[cfg(test)]
-mod transaction_test;
 pub mod write_api_error;
 pub mod write_api_result;
