+++ crates/papyrus_sync/src/sources/central.rs
@@ -1,3 +1,6 @@
+#[cfg(test)]
+#[path = "central_test.rs"]
+mod central_test;
 mod state_update_stream;
 
 use std::collections::{BTreeMap, HashMap};
