+++ crates/papyrus_sync/src/sources/mod.rs
@@ -2,5 +2,3 @@ pub mod base_layer;
 pub mod central;
 #[cfg(test)]
 mod central_sync_test;
-#[cfg(test)]
-mod central_test;
