+++ crates/starknet_client/src/reader/objects/mod.rs
@@ -1,7 +1,3 @@
 pub mod block;
-#[cfg(test)]
-mod block_test;
 pub mod state;
 pub mod transaction;
-#[cfg(test)]
-mod transaction_test;
