+++ crates/starknet_client/src/reader/objects/transaction_test.rs
@@ -1,7 +1,7 @@
 use assert::{assert_err, assert_ok};
 use assert_matches::assert_matches;
 
-use super::transaction::{Transaction, TransactionReceipt};
+use super::{Transaction, TransactionReceipt};
 use crate::test_utils::read_resource::read_resource_file;
 
 #[test]
