+++ crates/starknet_client/src/writer/objects/mod.rs
@@ -1,8 +1,4 @@
 pub mod response;
-#[cfg(test)]
-mod response_test;
 #[cfg(any(feature = "testing", test))]
 pub mod test_utils;
 pub mod transaction;
-#[cfg(test)]
-mod transaction_test;
