+++ crates/starknet_client/src/writer/objects/transaction.rs
@@ -6,6 +6,10 @@
 //!
 //! [`Starknet`]: https://starknet.io/
 
+#[cfg(test)]
+#[path = "transaction_test.rs"]
+mod transaction_test;
+
 use std::collections::HashMap;
 
 use serde::{Deserialize, Serialize};
